Turn off echoing
termios is a struct that we can use to modify how the terminal behaves.
We can get the attributes and read them into a struct using tcgetattr
and then we can turn echo off by using the c_lflags (local flags). Since
ECHO is a bitflag we can bitwise-NOT the operator and then bitwise-AND
to turn off the bit for ECHO. This is a common way to flip bits.