Disable raw mode at exit
By turning off ECHO and upon exit, the user will not receive any output
from typing. This is because we did not disable raw mode upon exiting.
We do that by creating a function and atexit() we pass in that func and
reset our terminal. TCSAFLUSH option will flush any unread output when
the settings are reset.